requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==2.3.7
Pillow==10.0.1
boto3==1.29.7
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.20.1
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-flask==1.3.0
black==23.9.1
flake8==6.1.0